backend/cms/serializers.py
from rest_framework import serializers

from cms.models import Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ["id", "title", "description", "price", "get_discounted_price"]
